chore: theme docs-only stories
