+++ .storybook/manager.js
@@ -1,8 +1,5 @@
 import { addons } from "@storybook/addons"
-import { create, themes } from "@storybook/theming"
-import { getTheme } from "../src"
-
-const theme = getTheme()
+import { storybookThemeLight } from "./theming"
 
 addons.setConfig({
   isFullscreen: false,
@@ -13,14 +10,5 @@ addons.setConfig({
   enableShortcuts: true,
   isToolshown: true,
   selectedPanel: undefined,
-  theme: create({
-    ...themes.normal,
-    base: `light`,
-    fontBase: theme.fonts.body,
-    fontCode: theme.fonts.monospace,
-    colorSecondary: theme.colors.gatsby,
-    brandTitle: `Gatsby Interface`,
-    barTextColor: theme.colors.grey[70],
-    barSelectedColor: theme.colors.purple[50],
-  }),
+  theme: storybookThemeLight,
 })
