+++ __stories__/about/Intro.stories.mdx
@@ -1,7 +1,9 @@
 import { Meta } from '@storybook/addon-docs/blocks';
-import { Heading, Text } from '../../src'
+import { MdxDocsTheme } from '../../.storybook/theming'
 
 <Meta title="About Gatsby Interface/Intro" />
+
+<MdxDocsTheme />
 
 # Intro
