+++ __stories__/about/Usage.stories.mdx
@@ -1,5 +1,5 @@
-import { Meta } from '@storybook/addon-docs/blocks';
-import { Heading, Text } from '../../src'
+import { Meta, Title } from '@storybook/addon-docs/blocks';
+import { MdxDocsTheme } from '../../.storybook/theming'
 import pkg from '../../package.json'
 
 export const peerDepsList = Object.keys(pkg.peerDependencies)
@@ -13,7 +13,7 @@ export const peerDepsList = Object.keys(pkg.peerDependencies)
 
 <Meta title="About Gatsby Interface/Usage" />
 
-
+<MdxDocsTheme />
 
 # How to use Gatsby Interface
